requirements.txt
pillow
pycryptodome~=3.9.0
opencv-python~=4.2.0
numpy~=1.18.0
matplotlib~=3.1.0